+ Response + HTTP methods HEAD, OPTIONS, PUT, DELETE, TRACE * Dep upgrades * Fixes

+++ app/src/main/java/us/huseli/retaintheme/request/HttpContentRange.kt
@@ -5,7 +5,7 @@ data class HttpContentRange(
     val unit: String,
     val rangeStart: Int,
     val rangeEnd: Int,
-    val size: Int? = null,
+    val size: Long? = null,
 )
 
 fun String.parseContentRange(): HttpContentRange? =
@@ -14,6 +14,6 @@ fun String.parseContentRange(): HttpContentRange? =
             unit = it[1],
             rangeStart = it[2].toInt(),
             rangeEnd = it[3].toInt(),
-            size = if (it[4] == "*") null else it[4].toInt(),
+            size = if (it[4] == "*") null else it[4].toLong(),
         )
     }
